backend/pyproject.toml
[project]
name = "openinstrument-backend"
version = "0.1.0"
description = "OpenInstrument - AI-Driven Instrumentation Design Platform Backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.0",
    "djangorestframework>=3.14",
    "django-cors-headers>=4.3",
    "psycopg[binary]>=3.1",
    "python-dotenv>=1.0",
    "djangorestframework-simplejwt>=5.3",
    "drf-spectacular>=0.27",
    "redis>=5.0",
    "django-redis>=5.4",
    "django-mptt>=0.14",
    "django-filter>=24.0",
    "jsonschema>=4.20",
    "pillow>=12.0.0",
    "django-tenants>=3.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-django>=4.5",
    "black>=23.0",
    "ruff>=0.1",
    "mypy>=1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["config", "apps"]

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "N", "W"]

[tool.black]
line-length = 88

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
